+++ sdmx-dl-provider-ri/src/main/java/internal/util/http/HttpResponseException.java
@@ -1,6 +1,7 @@
 package internal.util.http;
 
 import lombok.NonNull;
+import org.checkerframework.checker.nullness.qual.Nullable;
 
 import java.io.IOException;
 import java.util.Collections;
@@ -14,11 +15,11 @@ public final class HttpResponseException extends IOException {
     private final String responseMessage;
     private final Map<String, List<String>> headerFields;
 
-    public HttpResponseException(int responseCode, @NonNull String responseMessage) {
+    public HttpResponseException(int responseCode, @Nullable String responseMessage) {
         this(responseCode, responseMessage, Collections.emptyMap());
     }
 
-    public HttpResponseException(int responseCode, @NonNull String responseMessage, @NonNull Map<String, List<String>> headerFields) {
+    public HttpResponseException(int responseCode, @Nullable String responseMessage, @NonNull Map<String, List<String>> headerFields) {
         super(responseCode + ": " + responseMessage);
         this.responseCode = responseCode;
         this.responseMessage = responseMessage;